refactor: Heavy AI assistent code cleanup and split into composables
